Don't try to delete pending changes if the tx doesn't have any

+++ Raven.Client.Lightweight/Document/RavenClientEnlistment.cs
@@ -108,7 +108,10 @@ namespace Raven.Client.Document
 
 				using (var machineStoreForApplication = IsolatedStorageFile.GetMachineStoreForDomain())
 				{
-					machineStoreForApplication.DeleteFile(TransactionRecoveryInformationFileName);
+					if (machineStoreForApplication.FileExists(TransactionRecoveryInformationFileName))
+					{
+						machineStoreForApplication.DeleteFile(TransactionRecoveryInformationFileName);
+					}
 				}
 			}
 			catch (Exception e)
